Comment out unworking code in song controller and song model, add concern for shared behavior between passage and lyric
