+++ app/controllers/songs_controller.rb
@@ -6,11 +6,11 @@ class SongsController < ApplicationController
     genre_id = params[:genre].to_i
     genre = Genre.find_by(id: genre_id)
 
-    emotion = # Passage.primary_emotion
-    attribute_map = Song.map_emotions(emotion)
+    # emotion = # Passage.primary_emotion
+    # attribute_map = Song.map_emotions(emotion)
 
     @song = Song.new
-    recommendations = @song.get_recommendations({ seed_genres: genre.categories }, attribute_map)
+    recommendations = @song.get_recommendations({ seed_genres: genre.categories })
 
     recommendations.each do |song|
       Song.create(artist_name: song["artist"], song_title: song["track"])
